requirements.txt
django==5.2.8
django-colorfield==0.14.0